+++ lib/src/Globals.java
@@ -45,7 +45,7 @@ public class Globals
 
   public static final int BLOCK_CHUNK_HEADER_DOWNLOAD_SIZE = 500;
 
-  public static final int ADDRESS_HISTORY_MAX_REPLY = 100000;
+  public static final int ADDRESS_HISTORY_MAX_REPLY = 10000;
 
   public static final long MINE_CHUNK_SIZE = 1024L*1024L*1024L;