+++ RELEASE-NOTES.md
@@ -6,6 +6,7 @@
 * Update seed nodes
 * Add remote address to Shackleton explorer logs
 * Switch Shackleton web handling over to new web framework
+* Added explorer APIs zone and /api/total_coins
 
 ## 1.7.0