+++ RELEASE-NOTES.md
@@ -6,6 +6,7 @@
 * Added explorer APIs zone and /api/total_coins
 * Lower max history replies
 * Add get_address_history rpc
+* SIP4 - Input value checking
 
 ## 1.7.1